OK-509: Järjestelty paremmin valintakoekutsuihin liittyviä tyypityksiä
